requirements.txt
Django==4.2.7
google-generativeai==0.3.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
requests==2.31.0
python-dotenv==1.0.0